juggler speed fix

+++ sparrow/src/Classes/SPJuggler.h
@@ -89,6 +89,6 @@
 
 /// The speed factor adjusts how fast a juggler's animatables run.
 /// For example, a speed factor of 2.0 means the juggler runs twice as fast.
-@property (nonatomic, assign) double speed;
+@property (nonatomic, assign) float speed;
 
 @end
